requirements.txt
requests
argparse
python-dotenv[cli]